First commit.  Large commit.  Too large.  In future, commits will be after every single passing test or single refactoring.
Code summary:
MainActivity is the app entry point.  It starts GuideActivity.

GuideActivity displays a guide.  It holds a DataManager to a request a guide and implements DataItemListener to receive the guide.

ParseObjectManager is a DataManager.  It manages the interaction between the LoaderManager and an asynchronous Loader.
It implements LoaderManager.LoaderCallbacks to receive results from loaders managed by the LoaderManager.
It then notifies the passed DataItemListener, in this case GuideActivity, with the result.

ParseObjectLoader is the AsyncTaskLoader managed by the LoaderManager.
It queries the Parse server in a background thread and delivers the result.
In general, a loader provides asynchronous loading of data and automatically reloads previous data on an activity configuration change.

Next steps:
Clean up object graph managed by ParseObjectManager as a DataManager.  Clarify interfaces.
Continue building out UI for a single guide in GuideActivity.