New GoogleGuideMapTest mocks the MapListener (e.g. TourGuide) and tests that it is notified when the GoogleMap is clicked. GoogleGuideMap now implements GoogleMap callback. TestActivity created to help obtain GoogleGuideMap's GoogleMap dependency when testing.
(GuideListener) TourGuide implements setGuide in addition to setPlace.  This way, an if statement is not needed inside of setPlace.
DebugGuideService and DebugGuideServiceTest reflect this contract change.

TourGuide now implements MapListener.
TourGuideTest checks that the new TourGuideFactory sets the TourGuide as the GuideMap's map listener.
DebugGuideActivity uses TourGuideFactory to set up the map listener.
GoogleGuideMap implements setMapListener as required by GuideMap interface.

Marker style now set with MarkerStyle enum in GuideMap.

Next step is to test and implement behavior of TourGuide to MapListener callbacks, such as onMapClick and onMarkerClick.