TourGuideTest checks that guide cards are set when the guide and places are set.
It checks that the right guide card is shown when a map marker is clicked, and checks that that marker is centered.

It checks that the guide cards are hidden when the map is clicked.

TourGuide implements this functionality.

GoogleGuideMapTest checks that the map listener is notified with the right information when a marker is clicked.

It also checks that the map listener is notified when the map is clicked.

GoogleGuideMap implements this functionality with a GuideMarker map to track information associated with each marker.

GuideCards defines an interface that GuideCardPager has stubbed out.

Next steps are to test and implement GuideCardPager.